autodb_project/requirements.txt
psycopg2
tqdm
pandas
python-dotenv
requests